Extract statements_partitioned_query.go
